COBOL libraries maily for temperature conversions.
For the time being, here there's one library which contains functions to convert a given temperature in any one of the four most common scales [Celsius, Kelvin, Fahrenheit and Rankine] to its equivalent in another one of these four scales. The library contains a master function as well which takes as arguments the given temperature and a one character string, being "C", "K", "F" or "R", and then returns the temperature in all the four aforementioned scales.